requirements.txt
mlflow==2.16.0
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0.2
